Extract shell tool call dispatcher
